+++ app/src/main/java/com/developers/sprintsync/core/components/track/data/model/Track.kt
@@ -1,12 +1,10 @@
 package com.developers.sprintsync.core.components.track.data.model
 
-import android.graphics.Bitmap
-
 data class Track(
     val id: Int,
     val timestamp: Long,
     val durationMillis: Long,
-    val distanceMeters: Int,
+    val distanceMeters: Float,
     val avgPace: Float,
     val bestPace: Float,
     val calories: Int,
@@ -18,7 +16,7 @@ data class Track(
                 id = 0,
                 timestamp = 0L,
                 durationMillis = 0L,
-                distanceMeters = 0,
+                distanceMeters = 0f,
                 avgPace = 0f,
                 bestPace = 0f,
                 calories = 0,
