+++ app/src/main/java/com/developers/sprintsync/core/components/track/presentation/indicator_formatters/DistanceUiFormatter.kt
@@ -6,7 +6,7 @@ import java.util.Locale
 
 object DistanceUiFormatter {
     fun format(
-        distanceInMeters: Int,
+        distanceInMeters: Float,
         uiDistanceUiPattern: DistanceUiPattern,
     ): String {
         val locale = Locale.getDefault()
