+++ app/src/main/java/com/developers/sprintsync/statistics/domain/chart/formatters/axis/DefaultSelectiveYAxisValueFormatter.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import com.developers.sprintsync.core.components.goal.data.model.Metric
 import com.developers.sprintsync.core.components.track.presentation.indicator_formatters.CaloriesFormatter
 import com.developers.sprintsync.core.components.track.presentation.indicator_formatters.DistanceUiFormatter
-import com.developers.sprintsync.core.components.track.presentation.indicator_formatters.DurationFormatter
 import com.developers.sprintsync.core.components.track.presentation.indicator_formatters.DistanceUiPattern
+import com.developers.sprintsync.core.components.track.presentation.indicator_formatters.DurationFormatter
 import com.developers.sprintsync.core.util.extension.approximatelyEquals
 import kotlin.math.abs
 
@@ -39,13 +39,7 @@ class DefaultSelectiveYAxisValueFormatter : SelectiveYAxisValueFormatter() {
         ) {
             when (metric) {
                 Metric.DURATION -> DurationFormatter.formatToMm(selectedValue.toLong(), true)
-
-                Metric.DISTANCE ->
-                    DistanceUiFormatter.format(
-                        selectedValue.toInt(),
-                        DistanceUiPattern.WITH_UNIT,
-                    )
-
+                Metric.DISTANCE -> DistanceUiFormatter.format(selectedValue, DistanceUiPattern.WITH_UNIT)
                 Metric.CALORIES -> CaloriesFormatter.formatCalories(selectedValue.toInt(), true)
             }
         } else {
